Fix windows build error
